requirements.txt
requests
python-whois
dnspython
rich
phonenumbers
Pillow
